Made TypicalString have a 32 bit string, rather than an 8 bit string. Added some UTF8 string to UTF32 string decoding.

+++ translate in c/TypicalString-access.c
@@ -22,7 +22,7 @@
 
 class TypicalString : public TypicalObject {
 public:
-	int8 *data;
+	int32 *data;
 	int32 length;
 	int32 capacity;
 };
